fix locale bug
